Create log
Added Log API
Adds logging functionality. This API enables you to create logs in a new file (at definable path) with the option to show or hide text on the console while logging. A couple of convenient extra functions are also included.
Taken and updated from my own personal apis. May be of limited use for us but can be used for debugging. Can be left in for the end user as well.
Only one log can be stored per program. Logs get overwritten if the program attempts to create a new log.
For ease while developing, if you turn logging off you can force the api to delete unneeded leftover logs.